refactor: add local apis to local controller
