+++ seosa/src/main/java/com/seosa/seosa/domain/auth/local/dto/request/SignupRequestDTO.java
@@ -1,11 +1,9 @@
-package com.seosa.seosa.domain.auth.local.dto;
+package com.seosa.seosa.domain.auth.local.dto.request;
 
-import com.seosa.seosa.domain.user.entity.UserRole;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -13,7 +11,7 @@ import lombok.Setter;
 @Setter
 @Getter
 @NoArgsConstructor
-public class SignupDTO {
+public class SignupRequestDTO {
 
     @NotBlank(message = "이메일은 필수 입력 값입니다.")
     @Email(message = "이메일 형식이 올바르지 않습니다.")
