+++ seosa/src/main/java/com/seosa/seosa/domain/auth/local/dto/response/SignupResponseDTO.java
@@ -1,0 +1,11 @@
+package com.seosa.seosa.domain.auth.local.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class SignupResponseDTO {
+    private String message;
+    private String email;
+}
